requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
streamlit>=1.18.0
assemblyai>=0.16.0
tavily-python>=0.2.3
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
gtts>=2.3.2
pydub>=0.25.1
